feat: consolidate samoyed and samoyed-hook into unified binary architecture
Implements AC 63.1 by consolidating dual-binary architecture into a single
samoyed binary with Hook and Init subcommands, achieving deployment
simplification while maintaining architectural boundaries.

Changes:
- Add Hook subcommand to main samoyed binary with complete hook execution logic
- Convert samoyed-hook to deprecation shim delegating to samoyed hook
- Update hook script generation to call exec samoyed hook instead of exec samoyed-hook
- Add -f/--force flag to Init subcommand for future migration support
- Maintain backward compatibility with deprecation warnings until Sept 1, 2025

All 154 tests pass with zero clippy warnings. Preserves security boundaries,
dependency injection patterns, and cross-platform compatibility.

Closes #72
